+++ kitchen-sink/pages/dynamic-route.vue
@@ -1,7 +1,7 @@
 <template>
   <f7-page>
     <f7-navbar back-link="Back" title="Dynamic Route" sliding></f7-navbar>
-
+    <f7-block inner>
     <ul>
       <li><b>Url:</b> {{$route.url}}</li>
       <li><b>Path:</b> {{$route.path}}</li>
@@ -16,10 +16,20 @@
           <li v-for="(value, key) in $route.query"><b>{{key}}:</b> {{value}}</li>
         </ul>
       </li>
-      <li><b>Route:</b> {{$route.route}}</li>
+      <li><b>Route:</b> {{$route.route.path}}</li>
     </ul>
+    </f7-block>
+    <f7-block inner>
+      <p>Route <b>params</b> are also passed as component props:</p>
+      <ul>
+        <li><b>id:</b> {{id}}</li>
+        <li><b>post_id:</b> {{post_id}}</li>
+      </ul>
+    </f7-block>
   </f7-page>
 </template>
 <script>
-  export default {}
+  export default {
+    props: ['id', 'post_id']
+  }
 </script>