+++ src/components/navbar.vue
@@ -8,7 +8,7 @@
   >
     <slot name="before-inner"></slot>
     <div class="navbar-inner">
-      <f7-nav-left v-if="backLink" :back-link="backLink" :sliding="sliding" :back-link-href="backLinkUrl || backLinkHref"></f7-nav-left>
+      <f7-nav-left v-if="backLink" :back-link="backLink" :sliding="sliding" @back-click="onBackClick" :back-link-href="backLinkUrl || backLinkHref"></f7-nav-left>
       <f7-nav-center v-if="title" :title="title" :sliding="sliding"></f7-nav-center>
       <slot></slot>
     </div>
@@ -44,13 +44,13 @@
       }
     },
     methods: {
-      hide: function () {
+      hide: function (animated) {
         if (!this.$f7) return;
-        return this.$f7.hideNavbar(this.$el);
+        return this.$f7.hideNavbar(this.$el, animated);
       },
-      show: function () {
+      show: function (animated) {
         if (!this.$f7) return;
-        return this.$f7.showNavbar(this.$el);
+        return this.$f7.showNavbar(this.$el, animated);
       },
       size: function () {
         if (!this.$f7 || this.$theme.material) return;
@@ -67,6 +67,10 @@
       },
       onBeforeRemove: function (e) {
         this.$emit('navbar:beforeremove', e);
+      },
+      onBackClick: function (e) {        
+        this.$emit('back-click', e);
+        this.$emit('click:back', e);
       }
     }
   }